Add support for pedantic mode in consumer creation (#264)
Introduced a `pedantic` option in `ConsumerApiOptions` to ensure stricter validation of consumer configuration during creation. Updated the `add` method in `jsmconsumer_api.ts` to handle this option, and added a test case to validate pedantic behavior. This enhances flexibility while enforcing stricter checks when needed.

Note the original API allowed for an optional action (a string), now it also allows a ConsumerApiOptions object that allows setting the action and whether it should be pedantic.

Signed-off-by: Alberto Ricart <alberto@synadia.com>